src/main/webapp/app/entities/company-user/delete/company-user-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { CompanyUserService } from '../service/company-user.service';

import { CompanyUserDeleteDialogComponent } from './company-user-delete-dialog.component';

describe('Component Tests', () => {
  describe('CompanyUser Management Delete Component', () => {
    let comp: CompanyUserDeleteDialogComponent;
    let fixture: ComponentFixture<CompanyUserDeleteDialogComponent>;
    let service: CompanyUserService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [CompanyUserDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(CompanyUserDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(CompanyUserDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(CompanyUserService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
